requirements.txt
azure-core>=1.0.0
azure-functions
azure-storage-blob>=12.0.0
azure-storage-queue>=12.0.0
reportlab